can bind multiple buttons
the action stored for each button is now an array of actions.  The UI element is now a div that we append spans into, rather than a span we set the html of.  Also added a clear button on the bindings popup that will clear all the bindings.  This doesn't clear the UI at the moment.